% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis-functions.R
\name{city_coords_from_op_str_map}
\alias{city_coords_from_op_str_map}
\title{City coordinates from name}
\usage{
city_coords_from_op_str_map(city_name)
}
\arguments{
\item{city_name}{Name to look for in Open Street Map}
}
\value{
A simple feature with the name and the coordinates of the centre of the city.
}
\description{
City coordinates from name
}
\examples{
alingsas <- city_coords_from_op_str_map("Alingsås")
}
