% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_similar_strings.R
\name{find_similar_strings}
\alias{find_similar_strings}
\title{Title}
\usage{
find_similar_strings(
  strings,
  max_dist = 3,
  ignore_case = TRUE,
  costs = c(ins = 1, del = 1, sub = 1)
)
}
\arguments{
\item{strings}{Strings to compare}

\item{max_dist}{How many characters that should minimum diff.}

\item{ignore_case}{Defaults to True.}

\item{costs}{Complicated ...}
}
\description{
Title
}
