% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe_map.R
\name{globe_map}
\alias{globe_map}
\title{Title}
\usage{
globe_map(
  lat,
  lon,
  col_land = "#e9d6bd",
  col_water = "#abc1b1",
  col_grid = "aliceblue",
  col_border = "grey45",
  alpha_border = 1,
  alpha_grid = 0.8,
  n_longs = NULL,
  n_lats = NULL,
  show_coordinate_text = F,
  bbox_input = NULL,
  custom_map = NULL,
  get_only_crs = NULL
)
}
\arguments{
\item{lat}{Latitude of centre point}

\item{lon}{Longitude of centre point}

\item{col_land}{Land colour}

\item{col_water}{Water colour}

\item{col_grid}{Grid colour}

\item{col_border}{Border colour}

\item{alpha_border}{Transparency of border}

\item{alpha_grid}{Transparency of grid}

\item{n_longs}{Number of longitudes to show}

\item{n_lats}{Number of latitudes to show}

\item{show_coordinate_text}{Logical, show coordinates figures?}

\item{bbox_input}{Boundary box if you want to display less}

\item{custom_map}{Custom globe map to project}

\item{get_only_crs}{Logical: Do you wish only to return the CRS of the projection?}
}
\value{
An orhtographic map of the world centred on the given coordinates.
}
\description{
Title
}
